Moved 'Link Lost' to overwrite widget row
Made the 'Link lost' status message more obvious (and permanent) by
writing to the Widget Row (which is no longer being updated because the
connection is lost).  Means that it's bleedin obvious in the morning if
the watch lost connection overnight (happened twice due to low phone
battery).  Existing status icon in top corner to small to spot when half
asleep ;)